Drop the redundant cast on the strtol results
strtol returns long and there is no strtoi, so the narrowing to int has to
happen somewhere. Spelling it as a static_cast changes nothing, the implicit
conversion is the same one, and nothing here builds with -Wconversion.

+++ src/GUI/GUIFont.cpp
@@ -17,7 +17,7 @@ static Color parseColorTag(const char *tag, const Color &defaultColor) {
     char color_code[20];
     strncpy(color_code, tag, 5);
     color_code[5] = 0;
-    int color16 = static_cast<int>(std::strtol(color_code, nullptr, 10));
+    int color16 = std::strtol(color_code, nullptr, 10);
     if (color16 == 0) {
         return defaultColor; // Back to default color.
     } else {
@@ -296,7 +296,7 @@ std::string GUIFont::WrapText(std::string_view inString, int width, int uX, bool
                 char digits[4];
                 strncpy(digits, &inString[i + 1], 3);
                 digits[3] = 0;
-                lineWidth = static_cast<int>(std::strtol(digits, nullptr, 10)) + uX;
+                lineWidth = std::strtol(digits, nullptr, 10) + uX;
                 i += 3;
                 break;
             }
@@ -393,7 +393,7 @@ void GUIFont::DrawText(const Recti &rect, Pointi position, Color defaultColor, s
             strncpy(Dest, &string_base[i + 1], 3);
             Dest[3] = 0;
             i += 3;
-            left_margin = static_cast<int>(std::strtol(Dest, nullptr, 10));
+            left_margin = std::strtol(Dest, nullptr, 10);
             out_x = position.x + rect.x + left_margin;
             break;
         case '\n': // New line.
@@ -414,7 +414,7 @@ void GUIFont::DrawText(const Recti &rect, Pointi position, Color defaultColor, s
             strncpy(Dest, &string_base[i + 1], 3);
             Dest[3] = 0;
             i += 3;
-            left_margin = static_cast<int>(std::strtol(Dest, nullptr, 10));
+            left_margin = std::strtol(Dest, nullptr, 10);
             out_x = rect.x + rect.w - 1 - GetLineWidth(&string_base[i]) - left_margin;
             out_y = position.y + rect.y;
             if (maxY != 0) {
@@ -532,7 +532,7 @@ std::string GUIFont::FitTwoFontStringInWindow(std::string_view inString, GUIFont
                 char digits[4];
                 strncpy(digits, &inString[i + 1], 3);
                 digits[3] = 0;
-                lineWidth = static_cast<int>(std::strtol(digits, nullptr, 10)) + x;
+                lineWidth = std::strtol(digits, nullptr, 10) + x;
                 i += 3;
                 break;
             }
